refactor(canister-core): dissolve url module into its two call sites
Each helper in url.rs was used by exactly one module (plus tests), so
move them next to their sole caller and make them private:

- url_encode -> asset.rs (used only by render_env_cookie)
- url_decode + UrlDecodeError -> serving.rs (used only by http_request)

Their direct unit tests move into module-local #[cfg(test)] blocks
(matching existing precedent). The env_cookie tests that used url_decode
were simulating the browser's decodeURIComponent, not exercising our
decoder, so they now use a local client_decode helper over
percent_decode_str instead.
